+++ Source/Core/Core/HW/GCMemcard/GCMemcard.cpp
@@ -2,7 +2,7 @@
 // Licensed under GPLv2+
 // Refer to the license.txt file included.
 
-#include "Core/HW/GCMemcard.h"
+#include "Core/HW/GCMemcard/GCMemcard.h"
 
 #include <algorithm>
 #include <cinttypes>
